Move eventsListSkeleton to own component file
