+++ src/components/EventsList/EventsList.js
@@ -6,42 +6,9 @@ import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import GlobalStyles from 'assets/GlobalStyles';
 import EventContext from 'utils/EventContext';
-import Skeleton from '@material-ui/lab/Skeleton';
+import EventsListSkeleton from './EventsListSkeleton';
 
-function CustomSkeleton({ classes }) {
-  return (
-    <div className={classes.eventLists__loadingContainer}>
-      <Skeleton variant='rect' width={210} height={118} borderRadius={4} />
-      <div className={classes.eventLists__loadingTextContainer}>
-        <div style={{ gridColumn: '1 / span 2' }}>
-          <Skeleton variant='text' />
-          <Skeleton variant='text' />
-        </div>
-        <Skeleton style={{ gridColumn: '1/ span 1' }} variant='text' />
-      </div>
-    </div>
-  );
-}
 const eventListStyles = makeStyles((theme) => ({
-  eventLists__loadingContainer: {
-    paddingTop: '10px',
-    display: 'grid',
-    gridTemplateColumns: 'repeat(3, 1fr)',
-    '& .MuiSkeleton-root': {
-      backgroundColor: `${GlobalStyles.darkTeal}`,
-      borderRadius: '4px',
-    },
-  },
-
-  eventLists__loadingTextContainer: {
-    gridColumn: '2/ span 2',
-    paddingLeft: '10px',
-
-    display: 'grid',
-    gridTemplateColumns: 'repeat(2, 1fr)',
-    alignContent: 'space-between',
-  },
-
   eventLists__container: {
     height: '90vh',
     overflowY: 'scroll',
@@ -116,8 +83,8 @@ function EventsList({ startDate, latLng }) {
   return (
     <div className={classes.eventLists__container}>
       {loadingEvents
-        ? [...Array(numberOfSkeletons)].map((e, i) => (
-            <CustomSkeleton key={i} classes={classes} />
+        ? [...Array(numberOfSkeletons)].map((e) => (
+            <EventsListSkeleton key={e} />
           ))
         : eventsData.events.map((event, index) => (
             <>
